@@ -3,41 +3,40 @@
   "version": "0.0.0",
   "private": false,
   "dependencies": {
-    "@ckeditor/ckeditor5-build-classic": "^18.0.0",
-    "@ckeditor/ckeditor5-react": "^2.1.0",
+    "@ckeditor/ckeditor5-build-classic": "^33.0.0",
+    "@ckeditor/ckeditor5-react": "^4.0.1",
     "@reach/router": "^1.3.3",
-    "@sentry/browser": "^5.15.4",
-    "@sentry/node": "^5.15.4",
+    "@sentry/browser": "^6.19.3",
+    "@sentry/node": "^6.19.3",
     "aws-api-gateway-client": "^0.3.3",
-    "aws-sdk": "^2.651.0",
-    "axios": "^0.19.2",
+    "aws-sdk": "^2.1106.0",
+    "axios": "^0.26.1",
     "bcryptjs": "^2.4.3",
-    "cookie-parser": "^1.4.5",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "csvtojson": "^2.0.10",
-    "date-fns": "^2.19.0",
-    "dotenv": "^8.2.0",
-    "downshift": "^5.0.5",
-    "express": "^4.17.1",
-    "express-jwt": "^5.3.1",
-    "express-jwt-permissions": "^1.3.3",
-    "helmet": "^3.22.0",
-    "html-react-parser": "^0.10.3",
+    "date-fns": "^2.28.0",
+    "dotenv": "^16.0.0",
+    "downshift": "^6.1.7",
+    "express": "^4.17.3",
+    "express-jwt": "^6.1.1",
+    "express-jwt-permissions": "^1.3.6",
+    "helmet": "^5.0.2",
+    "html-react-parser": "^1.4.10",
     "jsonwebtoken": "^8.5.1",
     "lodash": "^4.17.15",
-    "mongoose": "^5.9.7",
-    "mongoose-unique-validator": "^2.0.3",
-    "nodemailer": "^6.4.6",
-    "passport": "^0.4.1",
+    "mongoose": "^6.2.9",
+    "mongoose-unique-validator": "^3.0.0",
+    "nodemailer": "^6.7.3",
+    "passport": "^0.5.2",
     "passport-local": "^1.0.0",
-    "pm2": "^4.2.3",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-scripts": "^3.4.1",
-    "react-spring": "^8.0.27",
-    "slugify": "^1.4.0",
-    "styled-components": "^5.0.1",
-    "validator": "^13.0.0"
+    "pm2": "^5.2.0",
+    "react": "^18.0.0",
+    "react-dom": "^18.0.0",
+    "react-scripts": "^5.0.0",
+    "slugify": "^1.6.5",
+    "styled-components": "^5.3.5",
+    "validator": "^13.7.0"
   },
   "scripts": {
     "start": "concurrently -k -p \"[{name}]\" -n \"Client,TypeScript,Node\" -c \"yellow.bold,cyan.bold,green.bold\" \"npm:client:start\" \"npm:server:watch:ts\" \"npm:server:watch:node\"",
@@ -57,45 +56,45 @@
     "server:build": "cd server && tsc"
   },
   "devDependencies": {
+    "@faker-js/faker": "^6.1.1",
     "@types/bcryptjs": "^2.4.2",
-    "@types/body-parser": "^1.17.1",
+    "@types/body-parser": "^1.19.2",
     "@types/cookie-parser": "^1.4.2",
-    "@types/cors": "^2.8.6",
-    "@types/dotenv": "^6.1.1",
-    "@types/express": "^4.17.1",
-    "@types/express-jwt": "0.0.42",
-    "@types/faker": "^4.1.5",
-    "@types/helmet": "0.0.44",
-    "@types/jest": "^24.0.18",
-    "@types/jsonwebtoken": "^8.3.4",
-    "@types/lodash": "^4.14.141",
-    "@types/mongodb": "^3.3.3",
-    "@types/mongodb-memory-server": "^1.8.0",
-    "@types/mongoose": "^5.5.18",
-    "@types/mongoose-unique-validator": "^1.0.2",
-    "@types/node": "^12.7.8",
-    "@types/nodemailer": "^6.2.1",
-    "@types/passport": "^1.0.1",
-    "@types/passport-local": "^1.0.33",
-    "@types/reach__router": "^1.2.5",
-    "@types/react": "^17.0.3",
-    "@types/react-dom": "^16.9.1",
-    "@types/styled-components": "^4.1.19",
-    "@types/supertest": "^2.0.8",
-    "@types/validator": "^10.11.3",
-    "concurrently": "^4.1.2",
-    "core-js": "^3.2.1",
-    "faker": "^4.1.0",
-    "husky": "^3.0.7",
-    "jest": "^24.9.0",
-    "lint-staged": "^9.4.0",
-    "mongodb-memory-server": "^5.2.5",
-    "prettier": "^1.18.2",
-    "react-app-polyfill": "^1.0.3",
+    "@types/cors": "^2.8.12",
+    "@types/dotenv": "^8.2.0",
+    "@types/express": "^4.17.13",
+    "@types/express-jwt": "6.0.4",
+    "@types/faker": "^6.6.9",
+    "@types/helmet": "4.0.0",
+    "@types/jest": "^27.4.1",
+    "@types/jsonwebtoken": "^8.5.8",
+    "@types/lodash": "^4.14.181",
+    "@types/mongodb": "^4.0.7",
+    "@types/mongodb-memory-server": "^2.3.0",
+    "@types/mongoose": "^5.11.97",
+    "@types/mongoose-unique-validator": "^1.0.5",
+    "@types/node": "^17.0.23",
+    "@types/nodemailer": "^6.4.4",
+    "@types/passport": "^1.0.7",
+    "@types/passport-local": "^1.0.34",
+    "@types/reach__router": "^1.3.10",
+    "@types/react": "^17.0.43",
+    "@types/react-dom": "^17.0.14",
+    "@types/styled-components": "^5.1.24",
+    "@types/supertest": "^2.0.12",
+    "@types/validator": "^13.7.2",
+    "concurrently": "^7.1.0",
+    "core-js": "^3.21.1",
+    "husky": "^7.0.4",
+    "jest": "^27.5.1",
+    "lint-staged": "^12.3.7",
+    "mongodb-memory-server": "^8.4.2",
+    "prettier": "^2.6.2",
+    "react-app-polyfill": "^3.0.0",
     "react-testing-library": "^8.0.1",
-    "supertest": "^4.0.2",
-    "ts-jest": "^24.1.0",
-    "typescript": "^3.6.3"
+    "supertest": "^6.2.2",
+    "ts-jest": "^27.1.4",
+    "typescript": "^4.6.3"
   },
   "husky": {
     "hooks": {
