@@ -1,7 +1,7 @@
 // @ts-ignore
 import ClassicEditor from '@ckeditor/ckeditor5-build-classic'
 // @ts-ignore
-import CKEditor from '@ckeditor/ckeditor5-react'
+import { CKEditor } from '@ckeditor/ckeditor5-react'
 import React from 'react'
 import styled from 'styled-components/macro'
 
